pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bifrost_bsx_tools"
version = "0.0.4"
authors = [
  { name="Valentin Bauer", email="valentin.bauer@tuwien.ac.at" },
]
description = "Extract time series from Bifrost BSX files as pandas DataFrame."
readme = "README.md"
license = { file = "LICENSE.txt" }
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "bifrost_common_py",
    "pandas",
    "pyyaml",
]
keywords = ["bifrost", "bsx"]

[project.urls]
"Homepage" = "https://github.com/va-li/bifrost-bsx-tools"
"Bug Tracker" = "https://github.com/va-li/bifrost-bsx-tools/issues"
"Bifrost" = "http://bifrost.siemens.com"
